+++ src/main/java/ru/yandex/practicum/filmorate/service/db/FilmDbService.java
@@ -87,7 +87,7 @@ public class FilmDbService implements FilmService {
     }
 
     @Override
-    public Collection<Film> getPopularFilms(int count, Long genreId, Integer year) {
+    public Collection<Film> getPopularFilms(int count, Integer genreId, Integer year) {
         return filmStorage.getPopularFilms(count, genreId, year);
     }
